+++ var/httpd/htdocs/js/Core.Agent.CustomerSearch.js
@@ -434,7 +434,10 @@ Core.Agent.CustomerSearch = (function (TargetNS) {
     };
 
     function htmlDecode(Text){
-        return Text.replace(/&amp;/g, '&');
+        return Text.replace(/&amp;/g, '&')
+            .replace(/&quot;/g, '"')
+            .replace(/&lt;/g, '<')
+            .replace(/&gt;/g, '>');
     }
 
     /**
@@ -463,6 +466,9 @@ Core.Agent.CustomerSearch = (function (TargetNS) {
 
         if (CustomerValue === '') {
             return false;
+        }
+        else {
+            CustomerValue = htmlDecode(CustomerValue);
         }
 
         // check for duplicated entries